+++ rosariosis.sql
@@ -1053,9 +1053,9 @@ CREATE TABLE food_service_staff_accounts (
 --
 
 CREATE TABLE food_service_staff_transaction_items (
-    -- @since 11.2 FS transaction item ID references food_service_menu_items(menu_item_id)
     item_id integer NOT NULL,
     transaction_id integer NOT NULL REFERENCES food_service_transactions(transaction_id),
+    menu_item_id integer,
     amount numeric(9,2),
     short_name varchar(25),
     description varchar(50),
@@ -1064,7 +1064,7 @@ CREATE TABLE food_service_staff_transaction_items (
     PRIMARY KEY (item_id, transaction_id)
 );
 
-COMMENT ON COLUMN food_service_staff_transaction_items.item_id IS 'References food_service_menu_items(menu_item_id)';
+COMMENT ON COLUMN food_service_staff_transaction_items.menu_item_id IS 'References food_service_menu_items(menu_item_id)';
 
 
 --
@@ -1107,9 +1107,9 @@ CREATE TABLE food_service_student_accounts (
 --
 
 CREATE TABLE food_service_transaction_items (
-    -- @since 11.2 FS transaction item ID references food_service_menu_items(menu_item_id)
     item_id integer NOT NULL,
     transaction_id integer NOT NULL REFERENCES food_service_transactions(transaction_id),
+    menu_item_id integer,
     amount numeric(9,2),
     discount varchar(25),
     short_name varchar(25),
@@ -1119,7 +1119,7 @@ CREATE TABLE food_service_transaction_items (
     PRIMARY KEY (item_id, transaction_id)
 );
 
-COMMENT ON COLUMN food_service_transaction_items.item_id IS 'References food_service_menu_items(menu_item_id)';
+COMMENT ON COLUMN food_service_transaction_items.menu_item_id IS 'References food_service_menu_items(menu_item_id)';
 
 
 --