+++ rosariosis_mysql.sql
@@ -1022,10 +1022,10 @@ CREATE TABLE food_service_staff_accounts (
 --
 
 CREATE TABLE food_service_staff_transaction_items (
-    -- @since 11.2 FS transaction item ID references food_service_menu_items(menu_item_id)
-    item_id integer NOT NULL COMMENT 'References food_service_menu_items(menu_item_id)',
+    item_id integer NOT NULL,
     transaction_id integer NOT NULL,
     FOREIGN KEY (transaction_id) REFERENCES food_service_staff_transactions(transaction_id),
+    menu_item_id integer COMMENT 'References food_service_menu_items(menu_item_id)',
     amount numeric(9,2),
     short_name varchar(25),
     description varchar(50),
@@ -1077,10 +1077,10 @@ CREATE TABLE food_service_student_accounts (
 --
 
 CREATE TABLE food_service_transaction_items (
-    -- @since 11.2 FS transaction item ID references food_service_menu_items(menu_item_id)
-    item_id integer NOT NULL COMMENT 'References food_service_menu_items(menu_item_id)',
+    item_id integer NOT NULL,
     transaction_id integer NOT NULL,
     FOREIGN KEY (transaction_id) REFERENCES food_service_transactions(transaction_id),
+    menu_item_id integer COMMENT 'References food_service_menu_items(menu_item_id)',
     amount numeric(9,2),
     discount varchar(25),
     short_name varchar(25),