+++ CHANGES.md
@@ -9,6 +9,7 @@ Changes in 11.2.1
 - Remove start & end date params from GET form URL in Expenses.php, Incomes.php & StudentSummary.php
 - Fix SQL syntax error at or near "ND" in DeleteTransactionItem.fnc.php
 - SQL add FOREIGN KEY to transaction_id column in rosariosis.sql & rosariosis_mysql.sql
+- SQL Add MENU_ITEM_ID column to food_service_transaction_items & food_service_staff_transaction_items tables in Update.fnc.php, rosariosis.sql & rosariosis_mysql.sql
 
 Changes in 11.2
 ---------------